Reducer consolidated simulation type
